# Use sync.pool for keccak256 and sha256