Before
```

BenchmarkHash-8                  1246952              1097 ns/op             179 B/op          4 allocs/op
BenchmarkHashKeccak256-8          344100              3352 ns/op             515 B/op          4 allocs/op

```

After
```
BenchmarkHash-8                  1967601               513 ns/op              35 B/op          2 allocs/op
BenchmarkHashKeccak256-8          807060              2536 ns/op             515 B/op          4 allocs/op
```


Non-trivial improvement, yay!