@nisdas 2020-02-01
@prestonvanloon could we try this for sha256 too ? There might be quite a few gains using it there